+++ package.json
@@ -10,7 +10,7 @@
     "emotion": "^9.2.6",
     "emotion-server": "^9.2.6",
     "gatsby": "next",
-    "gatsby-plugin-emotion": "^1.1.17",
+    "gatsby-plugin-emotion": "next",
     "gatsby-plugin-feed": "next",
     "gatsby-plugin-google-analytics": "next",
     "gatsby-plugin-manifest": "next",
@@ -32,7 +32,7 @@
     "prop-types": "^15.6.2",
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
-    "'react-emotion'": "^9.2.6",
+    "react-emotion": "^9.2.6",
     "react-helmet": "^5.2.0",
     "react-icons": "^3.0.5",
     "react-lazy-load": "^3.0.13",
