+++ src/components/TwitterIcon.js
@@ -1,9 +1,12 @@
 import React from 'react'
-import Icon from './Icon'
-import FaTwitterSquare from 'react-icons/lib/fa/twitter-square'
 
-export default (props: {}) => (
-  <Icon href="https://www.twitter.com/_SeanGroff" target="_blank">
-    <FaTwitterSquare size={25} />
-  </Icon>
-)
+import Icon from './Icon'
+import { FaTwitter } from 'react-icons/fa'
+
+export default function TwitterIcon() {
+  return (
+    <Icon href="https://www.twitter.com/_SeanGroff" target="_blank">
+      <FaTwitter size={25} />
+    </Icon>
+  )
+}
