+++ src/pages/index.js
@@ -1,77 +1,26 @@
 import React from 'react'
-import PropTypes from 'prop-types'
-import { Link, graphql } from 'gatsby'
-import get from 'lodash/get'
-import Helmet from 'react-helmet'
+import styled from 'react-emotion'
 import 'normalize.css'
 
-import Bio from '../components/Bio'
-import Layout from '../components/layout'
-import { rhythm } from '../utils/typography'
+import Hero from '../components/Hero'
+import About from '../components/About'
+import Portfolio from '../components/Portfolio'
+import ContactMe from '../components/ContactMe'
+import Footer from '../components/Footer'
 
-class BlogIndex extends React.Component {
-  static propTypes = {
-    location: PropTypes.object.isRequired,
-  }
-  render() {
-    const siteTitle = get(this, 'props.data.site.siteMetadata.title')
-    const siteDescription = get(this, 'props.data.site.siteMetadata.description')
-    const posts = get(this, 'props.data.allMarkdownRemark.edges')
-
-    return (
-      <Layout location={this.props.location}>
-        <Helmet
-          htmlAttributes={{ lang: 'en' }}
-          meta={[{ name: 'description', content: siteDescription }]}
-          title={siteTitle}
-        />
-        <Bio />
-        {posts.map(({ node }) => {
-          const title = get(node, 'frontmatter.title') || node.fields.slug
-          return (
-            <div key={node.fields.slug}>
-              <h3
-                style={{
-                  marginBottom: rhythm(1 / 4),
-                }}
-              >
-                <Link style={{ boxShadow: 'none' }} to={node.fields.slug}>
-                  {title}
-                </Link>
-              </h3>
-              <small>{node.frontmatter.date}</small>
-              <p dangerouslySetInnerHTML={{ __html: node.excerpt }} />
-            </div>
-          )
-        })}
-      </Layout>
-    )
-  }
-}
-
-export default BlogIndex
-
-export const pageQuery = graphql`
-  query {
-    site {
-      siteMetadata {
-        title
-        description
-      }
-    }
-    allMarkdownRemark(sort: { fields: [frontmatter___date], order: DESC }) {
-      edges {
-        node {
-          excerpt
-          fields {
-            slug
-          }
-          frontmatter {
-            date(formatString: "DD MMMM, YYYY")
-            title
-          }
-        }
-      }
-    }
-  }
+const Container = styled.div`
+  width: 100%;
+  height: 100vh;
 `
+
+export default function App() {
+  return (
+    <Container>
+      <Hero />
+      <About />
+      <Portfolio />
+      <ContactMe />
+      <Footer />
+    </Container>
+  )
+}
