+++ src/utils/typography.js
@@ -2,7 +2,7 @@ import Typography from 'typography'
 import Wordpress2016 from 'typography-theme-wordpress-2016'
 
 Wordpress2016.overrideThemeStyles = () => ({
-  'a.gatsby-resp-image-link': {
+  a: {
     boxShadow: 'none',
   },
 })
